refactor(webui): modify webui.env and webui.sh scripts
The changes made to the webui.env and webui.sh scripts refactor the installation process of the stable-diffusion-webui project. The git clone command in webui.sh now uses the SDW_GITHUB_USERNAME and SDW_GITHUB_REPO variables instead of the SDW_REPO_URL variable. Additionally, lines in the script related to creating and activating a python venv have been commented out. The modifications aim to improve the stability of the installation process.